[dagster graphql] add capture_error indirection
